requirements.txt
pillow
numpy
sqlalchemy
onnxruntime
rembg